[FLINK-10912][rocksdb] Make user defined options factory configurable
Add new interface ConfigurableOptionsFactory.
Enable user defined options factory could be configurable.
Add more docs about this change in `large_state_tuning.md`.
