+++ flink-state-backends/flink-statebackend-rocksdb/src/main/java/org/apache/flink/contrib/streaming/state/RocksDBConfigurableOptions.java
@@ -30,11 +30,11 @@ import static org.rocksdb.CompactionStyle.LEVEL;
 import static org.rocksdb.CompactionStyle.UNIVERSAL;
 
 /**
- * This class contains the configuration options for the {@link ConfigurableOptionsFactory}.
+ * This class contains the configuration options for the {@link DefaultConfigurableOptionsFactory}.
  *
  * <p>If nothing specified, RocksDB's options would be configured by {@link PredefinedOptions} and user-defined {@link OptionsFactory}.
  *
- * <p>If some options has been specifically configured, a corresponding {@link ConfigurableOptionsFactory} would be created
+ * <p>If some options has been specifically configured, a corresponding {@link DefaultConfigurableOptionsFactory} would be created
  * and applied on top of {@link PredefinedOptions} except if a user-defined {@link OptionsFactory} overrides it.
  */
 public class RocksDBConfigurableOptions implements Serializable {
